backend/requirements.txt
# FastAPI server + LLM query layer deps (Jossue)
# Install:
#   python -m venv .venv && source .venv/bin/activate
#   pip install -r requirements.txt

# API framework
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6

# HTTP client (for K2 calls + misc)
httpx>=0.27

# LLM clients
anthropic>=0.40

# Env loading
python-dotenv>=1.0

# SMS delivery for Eragon agent (optional — if TWILIO_* env vars unset,
# the agent degrades to draft-only and reports sent=false).
twilio>=9.0
